REFACTOR: Cleaned filter component page
