+++ components/Filters/FilterPanel.tsx
@@ -5,9 +5,9 @@ import FilterCategory from "@/types/filters/FilterCategory";
 import Link from "next/link";
 import React, { useState } from "react";
 import { IoClose } from "react-icons/io5";
-import RadioButton from "../../Inputs/RadioButton";
-import HeadingFour from "../../Text/HeadingFour";
-import HeadingThree from "../../Text/HeadingThree";
+import RadioButton from "../Inputs/RadioButton";
+import HeadingFour from "../Text/HeadingFour";
+import HeadingThree from "../Text/HeadingThree";
 
 interface FilterOverlayProps {
   filterCategories: FilterCategory[];
